requirements.txt
matplotlib==3.9.1.post1
numpy==2.0.1
pandas==2.2.2
pyodbc==5.1.0
requests==2.32.3
seaborn==0.13.2
SQLAlchemy==2.0.32
python-dotenv==1.0.1